requirements.txt
# Data & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Exchange API
ccxt>=4.0

# Machine Learning & Indicators
pandas-ta>=0.3.14

# API & Environment
anthropic>=0.25
requests>=2.31
python-dotenv>=1.0

# Notifications
python-telegram-bot>=20.0

# Dashboard
flask>=3.0
gunicorn>=21.0
